Small improvement.

+++ main.cpp
@@ -16,7 +16,7 @@ int main() {
 
     double min = -10;
     double max = 10;
-    unsigned int temperature = 100000;
+    double temperature = 100000;
     unsigned int itterations = 100000;
 
     annealingMeth(funcBooth, randomNum, temperature, itterations, X, min, max);
