+++ optMethods.h
@@ -11,7 +11,7 @@
 #include <cmath>
 
 void annealingMeth(double (*)(std::vector<double>&), double (*)(double&, double&),
-                   unsigned int&, unsigned int&, std::vector<double>&,
+                   double&, unsigned int&, std::vector<double>&,
                    double&, double&);
 
 #endif //ANNEALINGOPTIMISATION_OPTMETHODS_H
